Fix routing deprecation
