+++ .travis.yml
@@ -13,7 +13,6 @@ env:
     - DEPENDENCIES='beta'
 
 php:
-    - 5.6
     - 7.0
     - 7.1
     - 7.2
