+++ composer.json
@@ -35,7 +35,7 @@
         "swiftmailer/swiftmailer": "^5.2.0|^6.0",
         "symfony/asset": "^2.3|^3.0|^4.0",
         "symfony/browser-kit": "^2.3|^3.0|^4.0",
-        "symfony/framework-bundle": "^2.8|^3.0|^4.0",
+        "symfony/framework-bundle": "^2.8|^3.2|^4.0",
         "symfony/property-access": "^2.3|^3.0|^4.0",
         "symfony/security-bundle": "^2.3|^3.0|^4.0",
         "symfony/swiftmailer-bundle": "^2.3.11|^3.0",
