+++ tests/Controller/UpdatePasswordTest.php
@@ -1,0 +1,50 @@
+<?php
+
+/*
+ * This file is part of the CoopTilleulsForgotPasswordBundle package.
+ *
+ * (c) Vincent Chalamon <vincent@les-tilleuls.coop>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Tests\ForgotPasswordBundle\Controller;
+
+use CoopTilleuls\ForgotPasswordBundle\Controller\UpdatePassword;
+use CoopTilleuls\ForgotPasswordBundle\Entity\AbstractPasswordToken;
+use CoopTilleuls\ForgotPasswordBundle\Manager\ForgotPasswordManager;
+use Prophecy\Prophecy\ObjectProphecy;
+use Symfony\Component\HttpFoundation\Response;
+
+/**
+ * @author Vincent Chalamon <vincentchalamon@gmail.com>
+ */
+final class UpdatePasswordTest extends \PHPUnit_Framework_TestCase
+{
+    /**
+     * @var ForgotPasswordManager|ObjectProphecy
+     */
+    private $managerMock;
+
+    /**
+     * @var AbstractPasswordToken|ObjectProphecy
+     */
+    private $tokenMock;
+
+    protected function setUp()
+    {
+        $this->managerMock = $this->prophesize(ForgotPasswordManager::class);
+        $this->tokenMock = $this->prophesize(AbstractPasswordToken::class);
+    }
+
+    public function testUpdatePasswordAction()
+    {
+        $this->managerMock->updatePassword($this->tokenMock->reveal(), 'bar')->shouldBeCalledTimes(1);
+        $controller = new UpdatePassword($this->managerMock->reveal());
+        $response = $controller($this->tokenMock->reveal(), 'bar');
+        $this->assertInstanceOf(Response::class, $response);
+        $this->assertEquals('', $response->getContent());
+        $this->assertEquals(204, $response->getStatusCode());
+    }
+}
